firedoc-theme(1): introduce nodegit and npm
This wants to fix #35 in a curve road.
FireDoc shouldn't use bash/bat which is strongly based on a special OS,
now to address the issue, we introduce some programatic replacements for
npm and git:

- npm
- nodegit

and then we also rewrite the logic of scripts/install-theme.sh into
bin/firedoc-install.js.

Also removed the bash script because we don't need it anymore.